added enum calss
