+++ src/main/java/frc/robot/commands/FollowTarget.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj.PIDSource;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.*;
-import frc.robot.Enums.Target;
+import frc.robot.enums.Target;
 import frc.robot.pidsources.VisionPIDSourceX;
 import frc.robot.pidsources.VisionPIDSourceY;
 import frc.robot.utils.Limelight.CamMode;
