+++ src/main/java/frc/robot/commands/SetLiftHeight.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.PidSettings;
 import frc.robot.Robot;
 import frc.robot.RobotConstants;
-import frc.robot.Enums.LiftHeights;
+import frc.robot.enums.LiftHeights;
 import frc.robot.pidsources.LiftEncoderPidSource;
 
 public class SetLiftHeight extends Command {
