+++ src/main/java/frc/robot/commands/VisionPID.java
@@ -5,10 +5,10 @@ import edu.wpi.first.wpilibj.PIDSource;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.buttons.Button;
 import edu.wpi.first.wpilibj.command.Command;
-import frc.robot.Enums.Target;
 import frc.robot.PidSettings;
 import frc.robot.Robot;
 import frc.robot.RobotConstants;
+import frc.robot.enums.Target;
 import frc.robot.pidsources.VisionPIDSourceX;
 import frc.robot.pidsources.VisionPIDSourceY;
 import frc.robot.utils.Limelight.CamMode;
