+++ src/main/java/frc/robot/enums/ScoreHeight.java
@@ -1,0 +1,8 @@
+package frc.robot.enums;
+
+/**
+ * Score height enum
+ */
+public enum ScoreHeight {
+    kLow, kMedium, kHigh, kMediumGrossman
+}
